+++ src/app/(contents)/main/allzonecard/pagination/Pagination.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../styles/abstracts/core.scss' as *;
+@use '@/styles/abstracts/_core.scss' as *;
 
 .paginationWrapper {
   :global(.mantine-Pagination-control) {
