+++ src/app/(contents)/main/allzonecard/pagination/Pagination.tsx
@@ -1,7 +1,8 @@
 import { Pagination as MantinePagination } from '@mantine/core';
 import '@mantine/core/styles.css';
-import styles from '../main.module.scss';
-import paginationStyles from './Pagination.module.scss';
+import styles from '../AllZoneCard.module.scss';
+import S from './Pagination.module.scss';
+
 interface PaginationProps {
   onChangePage: (page: number) => void;
   pageCount: number;
@@ -10,7 +11,7 @@ interface PaginationProps {
 
 export default function Pagination({ onChangePage, pageCount, defaultValue }: PaginationProps) {
   return (
-    <div className={`${styles.pagination} ${paginationStyles.paginationWrapper}`}>
+    <div className={`${styles.pagination} ${S.paginationWrapper}`}>
       <MantinePagination total={pageCount} defaultValue={defaultValue} onChange={onChangePage} />
     </div>
   );
