+++ src/app/(contents)/main/banner/Banner.tsx
@@ -1,6 +1,6 @@
 import S from './Banner.module.scss';
 import Image from 'next/image';
-import { Activity } from '@/api/activities';
+import { Activity } from '@/fetches/activities';
 
 interface BannerProps {
   bestActivity: Activity | null;
