+++ src/app/(contents)/main/bestzonecard/BestZoneCard.module.scss
@@ -1,5 +1,18 @@
 @use '@/styles/abstracts/_core.scss' as *;
 
+.bestExperienceContainer {
+  display: flex;
+  justify-content: space-between;
+
+  margin-top: 35px;
+  margin-bottom: 32px;
+}
+
+.experienceText {
+  @include text3XlBold;
+  color: $black;
+}
+
 .bestZoneCardContainer {
   display: flex;
   gap: 24px;
